Fix autocomplete being broken on .NET 4.x Simplify evaluator usage by rolling AutoCompleteProvider into ShellEvaluator
