--- Assets/UnityShell/Editor/Scripts/AutocompleteProvider.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 95d25df2c4144f4c92eaa123f89bec93
-timeCreated: 1529695502
