+++ Assets/UnityShell/Editor/Scripts/UnityShellEditorWindow.cs
@@ -1,8 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Threading;
+using Mono.CSharp;
 using UnityEditor;
 using UnityEngine;
+using Event = UnityEngine.Event;
 
 namespace UnityShell
 {
@@ -65,15 +66,16 @@ namespace UnityShell
 				textEditor.text = value;
 			}
 		}
-
-		[SerializeField]
-		ShellEvaluator evaluator;
+		
+#if NET_4_6 || NET_STANDARD_2_0
+		private Evaluator evaluator;
+#endif
 
 		[SerializeField]
 		AutocompleteBox autocompleteBox;
 
 		[SerializeField]
-		AutocompleteProvider autocompleteProvider;
+		ShellEvaluator shellEvaluator;
 
 		[SerializeField]
 		Vector2 scrollPos = Vector2.zero;
@@ -103,7 +105,7 @@ namespace UnityShell
 			ClearText();
 			requestFocusOnTextArea = true;
 
-			autocompleteProvider = new AutocompleteProvider();
+			shellEvaluator = new ShellEvaluator();
 			autocompleteBox = new AutocompleteBox();
 		}
 
@@ -117,7 +119,6 @@ namespace UnityShell
 
 		void OnEnable()
 		{
-			new Thread(() => evaluator = new ShellEvaluator()).Start();
 			ScheduleMoveCursorToEnd();
 			autocompleteBox.onConfirm += OnAutocompleteConfirm;
 			autocompleteBox.Clear();
@@ -215,7 +216,7 @@ namespace UnityShell
 					lastWord = input.Substring(lastWordIndex + 1);
 				}
 
-				autocompleteProvider.SetInput(lastWord);
+				shellEvaluator.SetInput(lastWord);
 			}
 		}
 
@@ -299,7 +300,7 @@ namespace UnityShell
 			}
 			EditorGUILayout.EndScrollView();
 
-			autocompleteBox.results = autocompleteProvider.completions;
+			autocompleteBox.results = shellEvaluator.completions;
 			var pos = textEditor.graphicalCursorPos;
 			var rect = new Rect(pos.x, pos.y, 300, 200);
 			rect.y += 34;
@@ -319,7 +320,7 @@ namespace UnityShell
 					textEditor.MoveTextEnd();
 					try
 					{
-						var result = evaluator.Evaluate(input);
+						var result = shellEvaluator.Evaluate(input);
 						Append(result);
 						inputHistory.Add(input);
 						positionInHistory = inputHistory.Count;
